# graphql-mesh-todo-service/todo.express.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Express Todo API",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
    "/expressapp/todos": {
      "get": {
        "tags": [
          "ExpressTodoAPI"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TodoExpress"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ExpressTodoAPI"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TodoExpress"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoExpress"
                }
              }
            }
          }
        }
      }
    },    
    "/expressapp/todos/{id}": {      
      "delete": {
        "tags": [
          "ExpressTodoAPI"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TodoExpress": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "completed": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}